Main_code_file
This program allows defining the number of input and output layers. 

Training data needs to be given as input to the program.

Program updates the values in the weight and bias matrices and saves the values in the form of matrices. These values can be used for machine vishion later on.